rust: Move trivial types out of websocket.rs
